Second pass of the scrolling area fix.  Now ported to the Paradice9 application, where the main output window uses it.

+++ odin/ansi/adapt_protocol.hpp
@@ -41,8 +41,8 @@ BOOST_FUSION_ADAPT_STRUCT(
 BOOST_FUSION_ADAPT_STRUCT(
     odin::ansi::mouse_report,
     (odin::u8, button_)
-    (odin::u8, x_position_)
-    (odin::u8, y_position_)
+    (odin::s32, x_position_)
+    (odin::s32, y_position_)
 )
 
 #endif
